Remove misleading error log
This was introduced in https://github.com/ros/actionlib/pull/43.

It is not actually correct - you can feasibly get feedback here before a new goal is confirmed. See `send_goal()`....

```
    def send_goal(self, goal, done_cb=None, active_cb=None, feedback_cb=None):
        # destroys the old goal handle
        self.stop_tracking_goal()

        ...

        self.gh = self.action_client.send_goal(goal, self._handle_transition, self._handle_feedback)
```
and of course it will take more time on top of this for the server to actually process the incoming goal and confirm it. Meantime, it may have sent us feedback messages.