Remove debug log statment in mock_simple_server.py
For some reason, removing this log statement fixes a test failure.

I can confirm with print statements that execution reaches the end of execute_cb(),
but the goal result does not seem to reach the action client. The SimpleClientFixture.just_succeed
test gets stuck waiting for the result.

Removing the log statement (or changing to a verbosity higher than debug), fixes the issue.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>